requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
Django==4.2
docutils==0.19
dotty-dict==1.3.1
fastapi==0.89.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
image==1.5.33
importlib-metadata==6.5.0
invoke==1.7.3
jaraco.classes==3.2.3
joblib==1.2.0
keyring==23.13.1
more-itertools==9.1.0
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.27.0
packaging==23.1
Pillow==9.5.0
pkginfo==1.9.6
postgrest==0.10.6
psycopg2==2.9.5
pydantic==1.10.4
pyee==9.1.0
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-ffmpeg==2.0.4
python-gitlab==3.14.0
python-multipart==0.0.6
python-semantic-release==7.33.2
pywin32-ctypes==0.2.0
readme-renderer==37.3
realtime==1.0.0
regex==2023.3.23
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
starlette==0.22.0
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
tenacity==8.2.1
tomlkit==0.11.7
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.20.0
Wave==0.0.2
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
zipp==3.15.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.3
moviepy==1.0.3